[Refactor] Refactor List clusters
