+++ openstack/cce/v3/clusters/requests.go
@@ -1,96 +1,11 @@
 package clusters
 
 import (
-	"reflect"
-
 	golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
 )
 
 var RequestOpts = golangsdk.RequestOpts{
 	MoreHeaders: map[string]string{"Content-Type": "application/json"},
-}
-
-// ListOpts allows the filtering of list data using given parameters.
-type ListOpts struct {
-	Name  string `json:"name"`
-	ID    string `json:"uuid"`
-	Type  string `json:"type"`
-	VpcID string `json:"vpc"`
-	Phase string `json:"phase"`
-}
-
-// List returns collection of clusters.
-func List(client *golangsdk.ServiceClient, opts ListOpts) ([]Clusters, error) {
-	var r ListResult
-	_, r.Err = client.Get(rootURL(client), &r.Body, &golangsdk.RequestOpts{
-		OkCodes:     []int{200},
-		MoreHeaders: RequestOpts.MoreHeaders, JSONBody: nil,
-	})
-
-	allClusters, err := r.ExtractClusters()
-	if err != nil {
-		return nil, err
-	}
-
-	return FilterClusters(allClusters, opts), nil
-}
-
-func FilterClusters(clusters []Clusters, opts ListOpts) []Clusters {
-
-	var refinedClusters []Clusters
-	var matched bool
-	m := map[string]FilterStruct{}
-
-	if opts.Name != "" {
-		m["Name"] = FilterStruct{Value: opts.Name, Driller: []string{"Metadata"}}
-	}
-	if opts.ID != "" {
-		m["Id"] = FilterStruct{Value: opts.ID, Driller: []string{"Metadata"}}
-	}
-	if opts.Type != "" {
-		m["Type"] = FilterStruct{Value: opts.Type, Driller: []string{"Spec"}}
-	}
-	if opts.VpcID != "" {
-		m["VpcId"] = FilterStruct{Value: opts.VpcID, Driller: []string{"Spec", "HostNetwork"}}
-	}
-	if opts.Phase != "" {
-		m["Phase"] = FilterStruct{Value: opts.Phase, Driller: []string{"Status"}}
-	}
-
-	if len(m) > 0 && len(clusters) > 0 {
-		for _, cluster := range clusters {
-			matched = true
-
-			for key, value := range m {
-				if sVal := GetStructNestedField(&cluster, key, value.Driller); !(sVal == value.Value) {
-					matched = false
-				}
-			}
-			if matched {
-				refinedClusters = append(refinedClusters, cluster)
-			}
-		}
-
-	} else {
-		refinedClusters = clusters
-	}
-
-	return refinedClusters
-}
-
-type FilterStruct struct {
-	Value   string
-	Driller []string
-}
-
-func GetStructNestedField(v *Clusters, field string, structDriller []string) string {
-	r := reflect.ValueOf(v)
-	for _, drillField := range structDriller {
-		f := reflect.Indirect(r).FieldByName(drillField).Interface()
-		r = reflect.ValueOf(f)
-	}
-	f1 := reflect.Indirect(r).FieldByName(field)
-	return f1.String()
 }
 
 type ExpirationOptsBuilder interface {
