+++ openstack/cce/v3/clusters/urls.go
@@ -1,16 +1,12 @@
 package clusters
 
-import "github.com/opentelekomcloud/gophertelekomcloud"
+import golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
 
 const (
 	rootPath     = "clusters"
 	certPath     = "clustercert"
 	masterIpPath = "mastereip"
 )
-
-func rootURL(client *golangsdk.ServiceClient) string {
-	return client.ServiceURL(rootPath)
-}
 
 func resourceURL(c *golangsdk.ServiceClient, id string) string {
 	return c.ServiceURL(rootPath, id)
